@@ -1,21 +1,28 @@
 {
   "scripts": {
-    "dev": "webpack-dev-server",
-    "bundle": "webpack"
+    "dev": "webpack-dev-server --mode development",
+    "bundle": "webpack --mode production"
   },
   "dependencies": {
-    "efreet": "^0.1.9"
+    "efreet": "^0.2.20"
   },
   "devDependencies": {
-    "@types/velocity-animate": "^1.2.32",
-    "awesome-typescript-loader": "^3.4.1",
-    "css-loader": "^0.28.7",
+    "awesome-typescript-loader": "^5.0.0",
+    "css-loader": "^0.28.11",
+    "exports-loader": "^0.7.0",
+    "file-loader": "^2.0.0",
+    "imports-loader": "^0.8.0",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.19.1",
+    "style-loader": "^0.21.0",
     "stylus": "^0.54.5",
+    "stylus-loader": "^3.0.2",
     "stylus-relative-loader": "^3.4.0",
-    "typescript": "^2.6.2",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.9.7"
-  }
+    "raw-loader": "^0.5.1",
+    "typescript": "^3.1.3",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.12.0",
+    "webpack-cli": "^3.0.4",
+    "webpack-dev-server": "^3.1.4"
+  },
+  "capabilities": ["app"]
 }
